+++ Source/Utility/MythForest/Component/Stream/StreamComponentModule.h
@@ -22,7 +22,7 @@ namespace PaintsNow {
 			// static int main(int argc, char* argv[]);
 
 		public:
-			void RequestNew(IScript::Request& request, const UShort3& dimension, uint32_t cacheCount);
+			void RequestNew(IScript::Request& request, const UShort3& dimension, uint16_t cacheCount);
 			void RequestSetStreamLoadHandler(IScript::Request& request, IScript::Delegate<StreamComponent> streamComponent, IScript::Request::Ref ref);
 			void RequestSetStreamUnloadHandler(IScript::Request& request, IScript::Delegate<StreamComponent> streamComponent, IScript::Request::Ref ref);
 		};
